+++ CHANGELOG.md
@@ -41,7 +41,7 @@
 
 ### Fixed
 
-- None
+- replace deprecated db:structure by db:schema (#801).
 
 ## 3.3.0 - 2020-09-16